+++ lib/src/widgets/options/mp_cs_option_widget.dart
@@ -3,6 +3,7 @@ import 'package:mapiah/main.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
+import 'package:mapiah/src/widgets/options/mp_option_type_being_edited_tracking_mixin.dart';
 import 'package:mapiah/src/controllers/types/mp_window_type.dart';
 import 'package:mapiah/src/elements/command_options/th_command_option.dart';
 import 'package:mapiah/src/elements/parts/th_cs_part.dart';
@@ -33,7 +34,8 @@ class MPCSOptionWidget extends StatefulWidget {
   State<MPCSOptionWidget> createState() => _MPCSOptionWidgetState();
 }
 
-class _MPCSOptionWidgetState extends State<MPCSOptionWidget> {
+class _MPCSOptionWidgetState extends State<MPCSOptionWidget>
+    with MPOptionTypeBeingEditedTrackingMixin<MPCSOptionWidget> {
   String _selectedChoice = '';
   String _currentValue = '';
   String _osgbMajor = '';