+++ lib/src/widgets/options/mp_distance_type_option_widget.dart
@@ -4,6 +4,7 @@ import 'package:mapiah/src/auxiliary/mp_text_to_user.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
+import 'package:mapiah/src/widgets/options/mp_option_type_being_edited_tracking_mixin.dart';
 import 'package:mapiah/src/controllers/types/mp_window_type.dart';
 import 'package:mapiah/src/elements/command_options/th_command_option.dart';
 import 'package:mapiah/src/generated/i18n/app_localizations.dart';
@@ -34,8 +35,8 @@ class MPDistanceTypeOptionWidget extends StatefulWidget {
       _MPDistanceTypeOptionWidgetState();
 }
 
-class _MPDistanceTypeOptionWidgetState
-    extends State<MPDistanceTypeOptionWidget> {
+class _MPDistanceTypeOptionWidgetState extends State<MPDistanceTypeOptionWidget>
+    with MPOptionTypeBeingEditedTrackingMixin<MPDistanceTypeOptionWidget> {
   late TextEditingController _distanceController;
   late String _selectedUnit;
   late String _selectedChoice;