+++ lib/src/widgets/options/mp_double_value_option_widget.dart
@@ -3,6 +3,7 @@ import 'package:mapiah/main.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
+import 'package:mapiah/src/widgets/options/mp_option_type_being_edited_tracking_mixin.dart';
 import 'package:mapiah/src/controllers/types/mp_window_type.dart';
 import 'package:mapiah/src/elements/command_options/th_command_option.dart';
 import 'package:mapiah/src/generated/i18n/app_localizations.dart';
@@ -33,7 +34,8 @@ class MPDoubleValueOptionWidget extends StatefulWidget {
       _MPDoubleValueOptionWidgetState();
 }
 
-class _MPDoubleValueOptionWidgetState extends State<MPDoubleValueOptionWidget> {
+class _MPDoubleValueOptionWidgetState extends State<MPDoubleValueOptionWidget>
+    with MPOptionTypeBeingEditedTrackingMixin<MPDoubleValueOptionWidget> {
   late final TH2FileEditController th2FileEditController;
   late TextEditingController _doubleController;
   late String _selectedChoice;