+++ lib/src/widgets/options/mp_orientation_option_widget.dart
@@ -3,6 +3,7 @@ import 'package:mapiah/main.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
+import 'package:mapiah/src/widgets/options/mp_option_type_being_edited_tracking_mixin.dart';
 import 'package:mapiah/src/controllers/types/mp_window_type.dart';
 import 'package:mapiah/src/elements/command_options/th_command_option.dart';
 import 'package:mapiah/src/elements/parts/th_double_part.dart';
@@ -34,7 +35,8 @@ class MPOrientationOptionWidget extends StatefulWidget {
       _MPOrientationOptionWidgetState();
 }
 
-class _MPOrientationOptionWidgetState extends State<MPOrientationOptionWidget> {
+class _MPOrientationOptionWidgetState extends State<MPOrientationOptionWidget>
+    with MPOptionTypeBeingEditedTrackingMixin<MPOrientationOptionWidget> {
   late final TH2FileEditController th2FileEditController;
   late TextEditingController _azimuthController;
   late String _selectedChoice;