+++ lib/src/widgets/options/mp_passage_height_option_widget.dart
@@ -4,6 +4,7 @@ import 'package:mapiah/src/auxiliary/mp_text_to_user.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
+import 'package:mapiah/src/widgets/options/mp_option_type_being_edited_tracking_mixin.dart';
 import 'package:mapiah/src/controllers/types/mp_window_type.dart';
 import 'package:mapiah/src/elements/command_options/th_command_option.dart';
 import 'package:mapiah/src/generated/i18n/app_localizations.dart';
@@ -35,7 +36,8 @@ class MPPassageHeightOptionWidget extends StatefulWidget {
 }
 
 class _MPPassageHeightOptionWidgetState
-    extends State<MPPassageHeightOptionWidget> {
+    extends State<MPPassageHeightOptionWidget>
+    with MPOptionTypeBeingEditedTrackingMixin<MPPassageHeightOptionWidget> {
   String _selectedChoice = '';
   late final TextEditingController _depthController = TextEditingController();
   late final TextEditingController _heightController = TextEditingController();