+++ lib/src/widgets/options/mp_text_type_option_widget.dart
@@ -3,6 +3,7 @@ import 'package:mapiah/main.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
+import 'package:mapiah/src/widgets/options/mp_option_type_being_edited_tracking_mixin.dart';
 import 'package:mapiah/src/controllers/types/mp_window_type.dart';
 import 'package:mapiah/src/elements/command_options/th_command_option.dart';
 import 'package:mapiah/src/elements/parts/th_string_part.dart';
@@ -33,7 +34,8 @@ class MPTextTypeOptionWidget extends StatefulWidget {
   State<MPTextTypeOptionWidget> createState() => _MPTextTypeOptionWidgetState();
 }
 
-class _MPTextTypeOptionWidgetState extends State<MPTextTypeOptionWidget> {
+class _MPTextTypeOptionWidgetState extends State<MPTextTypeOptionWidget>
+    with MPOptionTypeBeingEditedTrackingMixin<MPTextTypeOptionWidget> {
   late TextEditingController _textController;
   late String _selectedChoice;
   final FocusNode _textTextFieldFocusNode = FocusNode();