+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   * Applying dart fix --apply to all code.
   * Flutter upgrade to 3.41.1
   * flutter pub upgrade --major-versions
+  * Created TH2FileEditElementEditController.optionTypeBeingEdited.
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release
 * New features: